3/4/16 Update
With this update the steering/camera code is pretty much done and ready
for testing. The beginning of the functions to switch between manual
control mode and automatic mode are added and will be updated via
interrupts. Now for acceleration/deceleration. Expect to see a version
of this done in C++ soon as well.